Removed adaptive dependencies from AI.* projects. (#6066)
Removed adaptive projects and tests
Tweaked tests and classes to work with after removing adaptive projects.

+++ libraries/Microsoft.Bot.Builder.AI.QnA/Models/QnAMakerTraceInfo.cs
@@ -50,7 +50,7 @@ namespace Microsoft.Bot.Builder.AI.QnA
         /// The minimum score threshold, used to filter returned results.
         /// </value>
         [JsonProperty("scoreThreshold")]
-        public float ScoreThreshold { get; set; }
+        public double ScoreThreshold { get; set; }
 
         /// <summary>
         /// Gets or sets number of ranked results that are asked to be returned.
